Update trending articles page url
