+++ src/api/index.js
@@ -2,9 +2,9 @@ export * from './getArticle'
 export * from './getArticleMediaInfo'
 export * from './getArticleMediaList'
 export * from './getArticleSummary'
-export * from './getExperimentArticles'
 export * from './getExperimentConfig'
 export * from './getLanglinks'
 export * from './getRandomArticleTitle'
 export * from './getSuggestedArticles'
+export * from './getTrendingArticles'
 export * from './search'
