+++ src/utils/experiment.js
@@ -1,7 +1,5 @@
 const GROUP_STORAGE_KEY = '2021-KaiOS-app-homepage-content-suggestions'
 
-const USER_COUNTRY_STORAGE_KEY = 'user-country'
-
 const GROUPS = [
   'control',
   'trending-articles'
@@ -34,9 +32,3 @@ export const getExperiment = () => {
 }
 
 export const isTrendingArticlesGroup = () => getGroup() === GROUPS[1]
-
-export const setUserCountry = country => {
-  localStorage.setItem(USER_COUNTRY_STORAGE_KEY, country)
-}
-
-export const getUserCountry = () => localStorage.getItem(USER_COUNTRY_STORAGE_KEY)
