+++ src/utils/index.js
@@ -21,4 +21,5 @@ export * from './sendErrorLog'
 export * from './sendEvent'
 export * from './sendFeedback'
 export * from './skipArticleIntro'
+export * from './userCountry'
 export * from './viewport'
